requirements.txt
eclipse-zenoh==0.10.1rc0

